feat: initial support of wasm memory64 proposal
* loader
  * limits encoding changes
  * `MemType` missing index type setup
  * assuming default is 32 mode
  * changes related values to uint64_t
* validator
  * introduce helper `getItValueType` to do index type checking (index
    type is a new concept that introduced in the proposal, stands for
    indexing type for memory modes, memory64 use I64 and default is I32)
  * type check
    * store/load instructions
    * atomic instructions
    * v128 instructions
  * page count is possible up to 64-bit now
  * offset is 64bit now
* execution
  * share more code for 32/64-bit mode
  * update related memory instructions
    * instruction load/store and load/store many
    * instruction atomic
* driver: add option
* stack manager: move `popIndexType` that pop value based on index type
  (refers to above description)
* spec test
    * use `wasm-dev` branch to enable memory64 test suites
    * fix `wasmedgeExecutorCoreTests` by fixing the `grow` instruction
    * fix memory limit test (32-bit mode testing)

Later fixup

* fix bus error, overflow problem

  The root cause is combined from two trouble

  1. In executor check we only check the offset didn't overflow by using
     maximal of uint64_t to substract it
  2. In memory check we assume
     `const uint64_t AccessLen = Offset + Length`
     will not overflow!

  but the first check didn't ensure the second won't happen, thus, the
  error occurs.

* fix `f32.load`

  If `ValVariant` stores `uint32_t` then must use `uint32_t`, but I use
  `uint64_t` and cause the problem. `valToIndex` now will judge index type
  of memory64 mode, then use the correct type to expand `ValVariant`.

* fix windows build

+++ lib/llvm/compiler.cpp
@@ -3959,7 +3959,7 @@ public:
     Builder.createFence(LLVMAtomicOrderingSequentiallyConsistent);
   }
   void compileAtomicNotify(unsigned MemoryIndex,
-                           unsigned MemoryOffset) noexcept {
+                           uint64_t MemoryOffset) noexcept {
     auto Count = stackPop();
     auto Addr = Builder.createZExt(Stack.back(), Context.Int64Ty);
     if (MemoryOffset != 0) {
@@ -3976,7 +3976,7 @@ public:
                 {Context.Int32Ty, Context.Int32Ty, Context.Int32Ty}, false)),
         {LLContext.getInt32(MemoryIndex), Offset, Count}));
   }
-  void compileAtomicWait(unsigned MemoryIndex, unsigned MemoryOffset,
+  void compileAtomicWait(unsigned MemoryIndex, uint64_t MemoryOffset,
                          LLVM::Type TargetType, uint32_t BitWidth) noexcept {
     auto Timeout = stackPop();
     auto ExpectedValue = Builder.createZExtOrTrunc(stackPop(), Context.Int64Ty);
@@ -3998,7 +3998,7 @@ public:
         {LLContext.getInt32(MemoryIndex), Offset, ExpectedValue, Timeout,
          LLContext.getInt32(BitWidth)}));
   }
-  void compileAtomicLoad(unsigned MemoryIndex, unsigned MemoryOffset,
+  void compileAtomicLoad(unsigned MemoryIndex, uint64_t MemoryOffset,
                          unsigned Alignment, LLVM::Type IntType,
                          LLVM::Type TargetType, bool Signed = false) noexcept {
 
@@ -4022,7 +4022,7 @@ public:
       Stack.back() = Builder.createZExt(Load, IntType);
     }
   }
-  void compileAtomicStore(unsigned MemoryIndex, unsigned MemoryOffset,
+  void compileAtomicStore(unsigned MemoryIndex, uint64_t MemoryOffset,
                           unsigned Alignment, LLVM::Type, LLVM::Type TargetType,
                           bool Signed = false) noexcept {
     auto V = stackPop();
@@ -4046,7 +4046,7 @@ public:
     Store.setOrdering(LLVMAtomicOrderingSequentiallyConsistent);
   }
 
-  void compileAtomicRMWOp(unsigned MemoryIndex, unsigned MemoryOffset,
+  void compileAtomicRMWOp(unsigned MemoryIndex, uint64_t MemoryOffset,
                           [[maybe_unused]] unsigned Alignment,
                           LLVMAtomicRMWBinOp BinOp, LLVM::Type IntType,
                           LLVM::Type TargetType, bool Signed = false) noexcept {
@@ -4072,7 +4072,7 @@ public:
       Stack.back() = Builder.createZExt(Ret, IntType);
     }
   }
-  void compileAtomicCompareExchange(unsigned MemoryIndex, unsigned MemoryOffset,
+  void compileAtomicCompareExchange(unsigned MemoryIndex, uint64_t MemoryOffset,
                                     [[maybe_unused]] unsigned Alignment,
                                     LLVM::Type IntType, LLVM::Type TargetType,
                                     bool Signed = false) noexcept {
@@ -4700,7 +4700,7 @@ private:
     }
   }
 
-  void compileLoadOp(unsigned MemoryIndex, unsigned Offset, unsigned Alignment,
+  void compileLoadOp(unsigned MemoryIndex, uint64_t Offset, unsigned Alignment,
                      LLVM::Type LoadTy) noexcept {
     if constexpr (kForceUnalignment) {
       Alignment = 0;
@@ -4717,7 +4717,7 @@ private:
     LoadInst.setAlignment(1 << Alignment);
     stackPush(LoadInst);
   }
-  void compileLoadOp(unsigned MemoryIndex, unsigned Offset, unsigned Alignment,
+  void compileLoadOp(unsigned MemoryIndex, uint64_t Offset, unsigned Alignment,
                      LLVM::Type LoadTy, LLVM::Type ExtendTy,
                      bool Signed) noexcept {
     compileLoadOp(MemoryIndex, Offset, Alignment, LoadTy);
@@ -4727,24 +4727,24 @@ private:
       Stack.back() = Builder.createZExt(Stack.back(), ExtendTy);
     }
   }
-  void compileVectorLoadOp(unsigned MemoryIndex, unsigned Offset,
+  void compileVectorLoadOp(unsigned MemoryIndex, uint64_t Offset,
                            unsigned Alignment, LLVM::Type LoadTy) noexcept {
     compileLoadOp(MemoryIndex, Offset, Alignment, LoadTy);
     Stack.back() = Builder.createBitCast(Stack.back(), Context.Int64x2Ty);
   }
-  void compileVectorLoadOp(unsigned MemoryIndex, unsigned Offset,
+  void compileVectorLoadOp(unsigned MemoryIndex, uint64_t Offset,
                            unsigned Alignment, LLVM::Type LoadTy,
                            LLVM::Type ExtendTy, bool Signed) noexcept {
     compileLoadOp(MemoryIndex, Offset, Alignment, LoadTy, ExtendTy, Signed);
     Stack.back() = Builder.createBitCast(Stack.back(), Context.Int64x2Ty);
   }
-  void compileSplatLoadOp(unsigned MemoryIndex, unsigned Offset,
+  void compileSplatLoadOp(unsigned MemoryIndex, uint64_t Offset,
                           unsigned Alignment, LLVM::Type LoadTy,
                           LLVM::Type VectorTy) noexcept {
     compileLoadOp(MemoryIndex, Offset, Alignment, LoadTy);
     compileSplatOp(VectorTy);
   }
-  void compileLoadLaneOp(unsigned MemoryIndex, unsigned Offset,
+  void compileLoadLaneOp(unsigned MemoryIndex, uint64_t Offset,
                          unsigned Alignment, unsigned Index, LLVM::Type LoadTy,
                          LLVM::Type VectorTy) noexcept {
     auto Vector = stackPop();
@@ -4755,7 +4755,7 @@ private:
                                     Value, LLContext.getInt64(Index)),
         Context.Int64x2Ty);
   }
-  void compileStoreOp(unsigned MemoryIndex, unsigned Offset, unsigned Alignment,
+  void compileStoreOp(unsigned MemoryIndex, uint64_t Offset, unsigned Alignment,
                       LLVM::Type LoadTy, bool Trunc = false,
                       bool BitCast = false) noexcept {
     if constexpr (kForceUnalignment) {
